+++ email/src/main/kotlin/com/noonoo/email/application/port/input/EmailSignUpUseCase.kt
@@ -2,6 +2,6 @@ package com.noonoo.email.application.port.input
 
 import com.noonoo.user.domain.collection.SignUpEmailSendMessage
 
-interface RegisterEmailUseCase {
+interface EmailSignUpUseCase {
     fun signUpEmailSend(signUpEmailSendMessage: SignUpEmailSendMessage)
 }
