+++ user/src/main/kotlin/com/noonoo/user/application/port/input/MemberSignUpUseCase.kt
@@ -2,7 +2,7 @@ package com.noonoo.user.application.port.input
 
 import com.noonoo.user.application.port.input.command.MemberSignUpCommand
 
-interface MemberAuthUseCase {
+interface MemberSignUpUseCase {
     fun signUp(memberSignUpCommand: MemberSignUpCommand)
 
     fun verifyEmail(token: String)
