+++ user/src/main/kotlin/com/noonoo/user/application/port/output/EmailSendEventCommandPort.kt
@@ -3,7 +3,7 @@ package com.noonoo.user.application.port.output
 import com.noonoo.user.domain.collection.EmailSendEventCollection
 import com.noonoo.user.domain.collection.SignUpEmailSendMessage
 
-interface EmailEventPort {
+interface EmailSendEventCommandPort {
     fun save(emailSendEventCollection: EmailSendEventCollection)
 
     fun sendEmailEvent(signUpEmailSendMessage: SignUpEmailSendMessage)
